disable reloadOnSearch for work packages list
This prevents reinstantiating a controller on changes to the query
params of a url.

In this case, we can now use $state.go instead of working around it via
$location.

We still have to prevent the suburi from being included twice, though.
Because $state (and $location) for that matter will append the sub-uri
anyway, we can't add it ourselves or we will end up with it being
prepended twice.